requirements.txt
# requirements we need

# libary for mqqt
paho-mqtt
